Colocate all moderation functionality into moderation service, combine from admin and repo services
