+++ packages/pds/src/services/index.ts
@@ -7,7 +7,7 @@ import { AuthService } from './auth'
 import { FeedService } from './feed'
 import { RecordService } from './record'
 import { RepoService } from './repo'
-import { AdminService } from './admin'
+import { ModerationService } from './moderation'
 
 export function createServices(resources: {
   messageQueue: MessageQueue
@@ -21,7 +21,7 @@ export function createServices(resources: {
     feed: FeedService.creator(imgUriBuilder),
     record: RecordService.creator(messageQueue),
     repo: RepoService.creator(messageQueue, blobstore),
-    admin: AdminService.creator(),
+    moderation: ModerationService.creator(messageQueue, blobstore),
   }
 }
 
@@ -31,7 +31,7 @@ export type Services = {
   feed: FromDb<FeedService>
   record: FromDb<RecordService>
   repo: FromDb<RepoService>
-  admin: FromDb<AdminService>
+  moderation: FromDb<ModerationService>
 }
 
 type FromDb<T> = (db: Database) => T
